requirements.txt
streamlit
torch
numpy
opencv-python
rasterio
pillow
openai